frontend/requirements.txt
streamlit
requests
Pillow
streamlit-cropper
